+++ packages/blade/private/server/worker/index.ts
@@ -9,6 +9,7 @@ import { router as projectRouter, triggers as triggerList } from 'server-list';
 
 import type { ServerContext } from '@/private/server/context';
 import { getWaitUntil, runQueries, toDashCase } from '@/private/server/utils/data';
+import { polyfillCompressionStream } from '@/private/server/utils/polyfills';
 import {
   getRequestGeoLocation,
   getRequestLanguages,
@@ -55,31 +56,10 @@ if (
   import.meta.env.BLADE_ENV === 'production' &&
   import.meta.env.__BLADE_PROVIDER === 'edge-worker'
 ) {
-  // Bun doesn't support `CompressionStream` yet, so we need to polyfill it.
-  if (typeof Bun !== 'undefined') {
-    // We use dynamic imports to prevent `esbuild` from detecting and inlining them.
-    const prefix = 'node:';
-    const { Readable, Writable } = await import(`${prefix}stream`);
-    const zlib = await import(`${prefix}zlib`);
+  // Enable necessary polyfilly on unsupported runtimes.
+  polyfillCompressionStream();
 
-    const transformMap = {
-      deflate: zlib.createDeflate,
-      'deflate-raw': zlib.createDeflateRaw,
-      gzip: zlib.createGzip,
-    };
-
-    globalThis.CompressionStream ??= class CompressionStream {
-      readable: ReadableStream;
-      writable: WritableStream;
-
-      constructor(format: keyof typeof transformMap) {
-        const handle = transformMap[format]();
-        this.readable = Readable.toWeb(handle) as unknown as ReadableStream;
-        this.writable = Writable.toWeb(handle) as unknown as WritableStream;
-      }
-    };
-  }
-
+  // Enable the compression middleware.
   app.use(compress());
 }
